+++ src/components/modal/PostAlert.jsx
@@ -1,6 +1,31 @@
 import { useHistory } from "react-router-dom";
+import { BASE_URL } from "../constants/baseUrl";
+import axios from "axios";
 import "./modalAlert.css";
-function PostAlert() {
+
+function PostAlert(props) {
+    const deletePost = async () => {
+        // const url = BASE_URL + "/post/62d60b3a82fdcc712f4d4713";
+        const post_id = props.data.post.id;
+        const url = BASE_URL + `/post/${post_id}`;
+        const token = localStorage.getItem("token");
+
+        try {
+            const res = await axios(url, {
+                method: "DELETE",
+                headers: {
+                    Authorization: `Bearer ${token}`,
+                    "Content-type": "application/json",
+                },
+            });
+            alert("삭제되었습니다.");
+            console.log(res);
+        } catch (err) {
+            // 404페이지로 이동
+            console.log(err);
+        }
+    };
+
     let history = useHistory();
     return (
         <div className="ModalAlert">
@@ -14,7 +39,9 @@ function PostAlert() {
                 >
                     취소
                 </button>
-                <button className="modalAlertBtn delete">삭제</button>
+                <button className="modalAlertBtn delete" onClick={deletePost}>
+                    삭제
+                </button>
             </div>
         </div>
     );
